Refine timer Remove animations Fix logic

+++ Flagger/ViewModels/FlaggerViewModel.swift
@@ -22,7 +22,7 @@ class FlaggerViewModel: ObservableObject {
     @Published var state: FlaggerViewModelState = .none
     @Published var activeGameMode: GameMode?
     @Published var highScores = [GameMode: Int]()
-    @Published var bestTimes = [GameMode: Int]()
+    @Published var bestTimes = [GameMode: Double]()
     private let flagService = FlagService()
     
     @MainActor
